requirements.txt
crewai==0.70.1
langchain==0.1.20
openai==1.35.0
paddleocr==2.7.3
paddlepaddle==2.6.1
pillow==10.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
openpyxl==3.1.2
pandas==2.1.4
sqlalchemy==2.0.23
pymysql==1.1.0
redis==5.0.1
celery==5.3.4
pyodbc==5.0.1